+++ spec/support/grammar_matchers.rb
@@ -2,7 +2,23 @@ require 'rspec/expectations'
 
 RSpec::Matchers.define :not_match do |input|
   match do |grammar|
-    RubySpeech::GRXML::Matcher.new(grammar).match(input).is_a?(RubySpeech::GRXML::NoMatch)
+    @result = RubySpeech::GRXML::Matcher.new(grammar).match(input)
+    @result.is_a?(RubySpeech::GRXML::NoMatch)
+  end
+
+  failure_message_for_should do |grammar|
+    "expected #{grammar} to not match #{input}, but received a #{@result.class}"
+  end
+end
+
+RSpec::Matchers.define :potentially_match do |input|
+  match do |grammar|
+    @result = RubySpeech::GRXML::Matcher.new(grammar).match(input)
+    @result.is_a?(RubySpeech::GRXML::PotentialMatch)
+  end
+
+  failure_message_for_should do |grammar|
+    "expected #{grammar} to potentially match #{input}, but received a #{@result.class}"
   end
 end
 
@@ -23,10 +39,10 @@ RSpec::Matchers.define :match do |input|
   failure_message_for_should do |grammar|
     messages = []
     unless @result.is_a?(RubySpeech::GRXML::Match)
-      messages << "expected a match result, got a #{@result.class}"
+      messages << "expected #{grammar} to match, got a #{@result.class}"
     end
 
-    unless @result.interpretation == @interpretation
+    if @result.respond_to?(:interpretation) && @result.interpretation != @interpretation
       messages << %{expected interpretation to be "#{@interpretation}" but received "#{@result.interpretation}"}
     end
 
@@ -50,11 +66,11 @@ RSpec::Matchers.define :max_match do |input|
 
   failure_message_for_should do |grammar|
     messages = []
-    unless @result.is_a?(RubySpeech::GRXML::Match)
-      messages << "expected a match result, got a #{@result.class}"
+    unless @result.is_a?(RubySpeech::GRXML::MaxMatch)
+      messages << "expected #{grammar} to max-match, got a #{@result.class}"
     end
 
-    unless @result.interpretation == @interpretation
+    if @result.respond_to?(:interpretation) && @result.interpretation != @interpretation
       messages << %{expected interpretation to be "#{@interpretation}" but received "#{@result.interpretation}"}
     end